+++ packages/x-data-grid/src/DataGrid/index.ts
@@ -1,2 +1,1 @@
 export * from './DataGrid';
-export { DATA_GRID_PROPS_DEFAULT_VALUES } from './useDataGridProps';
