+++ packages/x-data-grid/src/constants/index.ts
@@ -1,3 +1,4 @@
 export * from './envConstants';
 export * from './localeTextConstants';
 export * from './gridClasses';
+export * from './dataGridPropsDefaultValues';
