+++ packages/x-data-grid/src/hooks/features/dimensions/useGridDimensions.ts
@@ -34,7 +34,7 @@ import {
 } from '../rows/gridRowsUtils';
 import { getTotalHeaderHeight } from '../columns/gridColumnsUtils';
 import { GridStateInitializer } from '../../utils/useGridInitializeState';
-import { DATA_GRID_PROPS_DEFAULT_VALUES } from '../../../DataGrid/useDataGridProps';
+import { DATA_GRID_PROPS_DEFAULT_VALUES } from '../../../constants/dataGridPropsDefaultValues';
 
 type RootProps = Pick<
   DataGridProcessedProps,
